// front/src/app/model/stop.ts
/**
 * Targomo API
 * This is a sample server Targomo server.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: ichaouia.amine@gmx.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Point } from './point';


export interface Stop {
    stop_id?: string;
    name?: string;
    people_off?: number;
    people_on?: number;
    prev_stop_id?: string;
    next_stop_id?: string;
    reach_pop_30_walk?: number;
    reach_pop_30_bike?: number;
    lat?: number;
    lng?: number;
    geom?: Point;
    color_on?: string;
    color_off?: string;
    color_bike?: string;
    color_walk?: string;
}